+++ CHANGELOG.md
@@ -5,7 +5,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## In progress
-- update create keyword to return created object
+
+## [0.8.8] - 2024-10-07
+### Added
+- update create keyword to return created object [#143](https://github.com/devopsspiral/KubeLibrary/pull/143) by [@aisonaku]
 
 ## [0.8.7] - 2023-11-19
 ### Added